Feat(Events/Listener): Implemented Spring Application Event mechanism.
Created the OperationEvent class, extending ApplicationEvent, to define a custom event payload.

Developed the OperationListener class, which implements ApplicationListener to act as the handler for the custom event.

Used the @Component annotation to register the listener as a Spring Bean, and utilized the @Override method to define the event handling logic.

Created the OperationController to trigger (publish) the OperationEvent, and successfully observed the corresponding execution in the listener.